Hoan thanh chuc nang filter tasks

+++ src/components/List.js
@@ -1,13 +1,36 @@
 import React, { Component } from 'react';
 import ListItem from './ListItem';
+import ListItemNone from './ListItemNone';
 
 class List extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            filterName : "",
+            filterStatus : -1   // -1: All Status   0: Deactive Status  1: Active Status
+        };
+    }
+
+    handleChange = (event) => {
+        var name = event.target.name;
+        var value = name === "filterStatus" ? parseInt(event.target.value) : event.target.value;
+        this.props.handleFilter(name === "filterName" ? value : this.state.filterName, name === "filterStatus" ? value : this.state.filterStatus);
+        this.setState({
+            [name] : value
+        });
+    }
 
     render() {
         var { tasks } = this.props;
-        var elem = tasks.map((task, index) => {
-            return <ListItem key= { task.id } index= { index } task= {task} deleteItem = { this.props.deleteItem } updateItem = { this.props.updateItem } />;
-        }); 
+        var { filterName } = this.state;
+        var elem;
+        if (tasks.length === 0) {
+            elem = <ListItemNone></ListItemNone>;
+        } else {
+            elem = tasks.map((task, index) => {
+                return <ListItem key= { task.id } index= { index } task= {task} deleteItem = { this.props.deleteItem } updateItem = { this.props.updateItem } />;
+            });
+        }
         return (
             <table className="table">
                 <thead>
@@ -21,11 +44,11 @@ class List extends Component {
                 <tbody>
                     <tr>
                         <td></td>
-                        <td><input type="text" /></td>
-                        <td><select name="state" id="state">
-                            <option value="all">Tất cả</option>
-                            <option value="kich-hoat">Kích hoạt</option>
-                            <option value="chua-kich-hoat">Chưa kích hoạt</option>
+                        <td><input type="text" name="filterName" value = { filterName } onChange = { this.handleChange } /></td>
+                        <td><select name="filterStatus" id="state" onChange = { this.handleChange } >
+                            <option value="-1">Tất cả</option>
+                            <option value="1">Kích hoạt</option>
+                            <option value="0">Chưa kích hoạt</option>
                         </select></td>
                         <td></td>
                     </tr>
